+++ src/lib/translators/mongo4.ts
@@ -1,7 +1,7 @@
 /** This module contains mongo specific translation operations */
 
 import { $$ } from '@/lib/helpers';
-import { ConvertStep } from '@/lib/steps';
+import { ConvertStep, ToDateStep } from '@/lib/steps';
 import { Mongo36Translator } from '@/lib/translators/mongo';
 
 type PropMap<T> = { [prop: string]: T };
@@ -30,7 +30,16 @@ function transformConvert(step: Readonly<ConvertStep>): MongoStep {
   return { $addFields: mongoAddFields };
 }
 
+/** transform a 'todate' step into corresponding mongo steps */
+function transformToDate(step: Readonly<ToDateStep>): MongoStep {
+  const dateFromString: MongoStep = { dateString: $$(step.column) };
+  if (step.format) {
+    dateFromString['format'] = step.format;
+  }
+  return { $addFields: { [step.column]: { $dateFromString: dateFromString } } };
+}
+
 export class Mongo40Translator extends Mongo36Translator {
   static label = 'Mongo 4.0';
 }
-Object.assign(Mongo40Translator.prototype, { convert: transformConvert });
+Object.assign(Mongo40Translator.prototype, { convert: transformConvert, todate: transformToDate });