+++ tests/unit/todate-step-form.spec.ts
@@ -2,12 +2,17 @@ import ToDateStepForm from '@/components/stepforms/ToDateStepForm.vue';
 
 import { BasicStepFormTestRunner } from './utils';
 
-describe('Connvert STring to Date Step Form', () => {
+describe('Convert String to Date Step Form', () => {
   const runner = new BasicStepFormTestRunner(ToDateStepForm, 'todate');
   runner.testInstantiate();
-  runner.testExpectedComponents({
-    'columnpicker-stub': 1,
-  });
+  runner.testExpectedComponents(
+    { 'columnpicker-stub': 1, 'inputtextwidget-stub': 0 },
+    { translator: 'mongo36' },
+  );
+  runner.testExpectedComponents(
+    { 'columnpicker-stub': 1, 'inputtextwidget-stub': 1 },
+    { translator: 'mongo40' },
+  );
 
   it('should update editedStep with the selected column at creation', () => {
     const initialState = {